JAMES-1902 Extracts in memory mailqueue guice bindings
